style(): made some small changes
