--- src/header/Logo.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-class Logo extends Component {
-  render () {
-    return (
-      <>
-        հայերեն անիմե ու մանգա
-      </>
-    )
-  }
-}
-
-export default Logo
