--- src/header/sass/Navigator.scss
@@ -1,76 +1,0 @@
-@import '../../sass/index';
-
-.navigator {
-
-    height: $menu-height;
-
-    position: relative;
-
-    .menu {
-
-        @include horizontal-menu;
-        
-        position: absolute;
-
-        bottom: 5px;
-
-        ul {
-            margin-bottom: 0;
-        }
-
-        div {
-
-            // make it a horizontal list
-            display: inline;
-            padding-bottom: 5px;
-
-        }  
-
-        li {
-            font-size: 25px;
-            font-weight: 500;
-            text-transform: uppercase;
-        }
-        
-    }
-
-    .theme-switch {
-
-        position: absolute;
-        bottom: 5px;
-        right: 5px;
-
-    }
-
-    .right-pane {
-
-        position: absolute;
-        right: 0;
-        
-        div {
-            display: inline-block;
-            padding: 10px;
-        }
-
-        .searchbar {
-
-            .ant-input-search {
-
-                width: 20vw;
-                // location
-                // position: absolute;
-                bottom: 0px;
-                right: 18.5vw;
-
-            }
-
-        }
-
-        .logo {
-
-            font-size: 20px;
-            width: 18vw;
-
-        }
-    }
-}
