+++ src/header/sass/Ribbon.scss
@@ -3,20 +3,26 @@
 .ribbon {
     
     @include horizontal-menu;
-    background-color: $black6;
     display: flow-root;
+    position: relative;
 
-    a {
+    border-bottom: 2px solid;
+    
+    li {
 
-        color: $color1;
+        text-transform: uppercase;
+        font-size: 20px;
 
     }
 
-    // .ribbon-manga {
-        
-    // }
+    .theme-switch {
 
-    // .ribbon-anime {
-        
-    // }
+        padding: 5px;
+
+        position: absolute;
+        bottom: 5px;
+        right: 5px;
+
+    }
+
 }
