+++ src/sass/index.scss
@@ -23,36 +23,32 @@ $color1: #1abc9c;
 $color2: #f27935;
 
 $left: 20px;
-$menu-height: 100px;
+$menu-height: 10px;
 
 @mixin base {
 
-  .navigator {
+  .ribbon {
 
     .menu {
 
-        div {
-        
-            &.selected {
+      a {
 
-                border-color: $color0;
+        &:hover {
+          color: $color0;
+        }
 
-                a {
-                    color: $color0;
-                }
-                
-            }
+      }
 
-        }  
+      .selected {
 
         a {
 
-          &:hover {
-                color: $color0;
-            }
+          color: $color0;
 
         }
         
+      }
+      
     }
 
   }
@@ -89,48 +85,21 @@ $menu-height: 100px;
 
   $outline-color: $black3;
 
-  .piccha {
+  .ribbon {
 
-    background-image: url('https://raw.githubusercontent.com/high-otaku/assets/master/background.png');
-
-    .info {
-
-      background-color: rgba($white0, 0.7);
-
-    }
-
-    h1 {
-      
-      color: $black8;
-
-    }
-
-  }
-
-  .navigator {
-    
-    background-color: rgba($black0, 0.7);
-    border-bottom: 2px solid $outline-color;
+    background-color: $white3;
+    border-color: $outline-color;
 
     .menu {
 
-        a {
-            color: $black6;
+      a {
 
-        }
-        
+        color: $black8;
+
+      }
+      
     }
 
-
-    .right-pane {
-
-        .logo {
-
-            background-color: $white0;
-
-        }
-    }
-    
   }
   
   .complete {
@@ -147,10 +116,6 @@ $menu-height: 100px;
 
     background-color: $white3;
     border-color: $outline-color;
-
-    &:hover {
-
-    }
 
   }
 
@@ -201,44 +166,19 @@ $menu-height: 100px;
 
   $outline-color: $black1;
 
-  .piccha {
+  .ribbon {
 
-    background-image: url('https://raw.githubusercontent.com/high-otaku/assets/master/background.png');
-
-    .info {
-
-      background-color: rgba($black7, 0.7);
-
-    }
-
-    h1 {
-      
-      color: $white0;
-      
-    }
-
-  }
-
-  .navigator {
-    
-    background-color: rgba($black6, 0.7);
-    border-bottom: 2px solid $outline-color;
+    background-color: $black5;
+    border-color: $outline-color;
 
     .menu {
 
-        a {
-            color: $white3;
-        }
-        
-    }
+      a {
 
-    .right-pane {
+        color: $white0;
 
-        .logo {
-
-            background-color: $black6;
-
-        }
+      }
+      
     }
 
   }
@@ -325,6 +265,8 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+
+  // background-image: url('https://raw.githubusercontent.com/high-otaku/assets/master/background.png');
 
   &.light-theme {
